requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
websockets==12.0
aiosqlite==0.19.0
